feat: Validação no admin
Validação para as datas de inicio e encerramento de avisos adicionada

+++ utilidades/admin/aviso_admin.py
@@ -2,6 +2,9 @@ from django.contrib import admin
 from django_object_actions import DjangoObjectActions
 
 from ..models import Aviso
+from ..forms import AvisoForm
+
+from datetime import datetime
 
 @admin.register(Aviso)
 class AvisoAdmin(DjangoObjectActions, admin.ModelAdmin):
@@ -17,9 +20,15 @@ class AvisoAdmin(DjangoObjectActions, admin.ModelAdmin):
     ]
     
     changelist_actions = [
-        
+        'remover_anuncios'
     ]
+    
+    form = AvisoForm
     
     def remover_anuncios(self, request, obj):
         for instance in obj:
-            instance.delete()
+            data_atual = datetime.strftime(datetime.now(), '%d/%m/%Y %H:%M:%S')
+            data_marcada = datetime.strftime(instance.data_de_encerramento, '%d/%m/%Y %H:%M:%S')
+            if data_atual >= data_marcada:
+                instance.delete()
+                self.message_user(request, 'Histórico de anúncio deletado com sucesso!')
